pyproject.toml
[project]
name = "momenta"
authors = [{name = "Mathieu Lamoureux"}]
version = "1.0"
dependencies = [
    "astropy",
    "cached_property",
    "datetime",
    "h5py",
    "healpy",
    "ligo.skymap==2.0.1",
    "matplotlib",
    "numpy==1.26.4",
    "pandas",
    "pyyaml",
    "scipy",
    "typing",
    "ultranest==4.3.1"
]

[project.optional-dependencies]
dev = [
    "black",
    "pytest",
    "pytest-cov",
    "pytest-flake8",
    "pytest-pylint",
    "pytest-watch",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_also = [
	"pass",
	"pragma: no cover",
	"raise AssertionError",
	"raise NotImplementedError",
	"if 0:",
	"if False:",
]

[tool.coverage.html]
directory = "coverage_html_report"

[tool.black]
line-length = 150
